+++ src/main/store/store.ts
@@ -81,6 +81,7 @@ export const DEFAULT_SETTINGS: SettingsData = {
   server: {
     enabled: false,
     readonly: false,
+    readonlyExtensionUi: true,
     basicAuth: {
       enabled: false,
       username: '',